refactor: consolidate data management
