+++ OPTIMIZATION_SUMMARY_V3.md
@@ -6,7 +6,7 @@
 
 ## 🚀 核心优化内容
 
-### 1. 数据统一管理系统 (DataManagerV3)
+### 1. 数据统一管理系统 (DataManager)
 
 **优化前：**
 - 数据分散在各个模块中
@@ -127,7 +127,7 @@ load_game_data()
 ### 1. 模块化设计
 ```
 xwe/core/
-├── data_manager_v3.py     # 数据管理
+├── data_manager.py        # 数据管理
 ├── formula_engine.py      # 公式引擎
 ├── cultivation_system.py  # 修炼系统
 ├── combat_system_v3.py    # 战斗系统
