+++ xwe/core/npc_system_v3.py
@@ -6,7 +6,7 @@ NPC系统优化实现
 import random
 import logging
 from typing import Dict, Any, List, Optional, Tuple
-from .data_manager_v3 import DM
+from .data_manager import DM
 from .formula_engine import formula_engine, evaluate_expression
 
 logger = logging.getLogger(__name__)
